Satellite_Property_Valuation/requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
jupyter
notebook
requests
python-dotenv
tqdm
Pillow
